find_house_proj/requirements.txt
asgiref==3.8.1
Django==4.2
pillow==10.4.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
